+++ helpers/ft_calloc.c
@@ -6,17 +6,17 @@
 /*   By: maakhan <maakhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/25 18:48:44 by maakhan           #+#    #+#             */
-/*   Updated: 2024/12/03 14:12:42 by maakhan          ###   ########.fr       */
+/*   Updated: 2024/12/24 13:08:47 by maakhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-void	*ft_calloc(size_t count, size_t size)
+void	*ft_calloc(int count, int size)
 {
 	void	*ptr;
 
-	if (size && count > SIZE_MAX / size)
+	if (size && (unsigned long)count > SIZE_MAX / size)
 		return (NULL);
 	ptr = malloc(count * size);
 	if (!ptr)
