+++ helpers/ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: maakhan <maakhan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/22 17:08:41 by maakhan           #+#    #+#             */
-/*   Updated: 2024/12/22 17:08:42 by maakhan          ###   ########.fr       */
+/*   Updated: 2024/12/24 12:36:51 by maakhan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, int len)
 {
 	void	*hold;
-	size_t	i;
+	int		i;
 
 	hold = b;
 	i = 0;
